refactor: model 패키지 domain으로 변경
