+++ src/main/java/domain/Car.java
@@ -1,4 +1,4 @@
-package model;
+package domain;
 
 import constant.ExceptionMessage;
